requirements.txt
openai
python-dotenv
whisper
rapidfuzz
ffmpeg-python
streamlit
pandas
ringcentral>=0.8.0